+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.156 (2023-12-5)
+
+- fix: remove all original files
+- feat: save project info to git
+
+## 0.1.155 (2023-12-5)
+
+- feat: new projects pages
+
 ## 0.1.154 (2023-12-1)
 
 - feat: runtime package